Updated the list of ignored symbols in partially linked file
 ### Summary of Changes ###

This commit makes it build and enables running the entry point for our
simulation. It updates the list of names that we don't want to export in
the partially linked file to avoid the conflict with the host
definitions. It adds the functions that should be implemented to have a
minimal build with console running:

uart_low_initialize()
get_console_sema()
putc()

Signed-off-by: Sebastian Ene <sebastian.ene07@gmail.com>