### Summary of Changes ###
Create a ring buffer that is shared between the host and the simulator
thread. This ring buffer will hold the UART characters as received in
the host thread. After we receive one character we send a signal to the
simulator thread to notify it about the new data. The data will be
receive on the simulator thread in the UART interrupt handler and it
will be copied in the lower half buffer.

Signed-off-by: Sebastian Ene <sebastian.ene07@gmail.com>