### Summary of Changes ###
The host thread which drives the external events simulation is started
before the Calypso OS simulator thread. Because of that, characters are
consumed from the stdin in the host thread loop and are sent to the
uninitialized uart lowerhalf driver in the simulation. Add a flag to
prevent stdin character consumption until the simulator thread opens the
simulated UART device.

Signed-off-by: Sebastian Ene <sebastian.ene07@gmail.com>