Verified working BLE Proximity application with softdevice on nrf52832
 ## Summary of changes

Update the README.md from nrf52832_softdevice to add instructions on the
softdevice library flash. Verified that the BLE proximity application
works without issues.

To flash the softdevice on the nrf52 we first need to clear the flash:

nrfjprog --eraseall
nrfjprog -f nrf52 --program s132_nrf52_1.0.0-3.alpha_softdevice.hex

The softdevice will be placed at the start of the flash. The Calypso OS
is placed after 0x1F000 in flash.

Signed-off-by: Sebastian Ene <sebastian.ene07@gmail.com>