Add rpipico build target and create skeleton
 ## Summary of changes

Build the pico-sdk as a separate static library and link Calypso OS with
this library. Build the 2nd stage bootloader and add padding bytes and
checksum. The checksum will be verified by the ROM bootloader.
The 2nd stage bootloader will initialize XIP and will jump in the Calypso OS
code at 0x1000100 (in XIP) where the vector table should reside.

Signed-off-by: dedsecn00b <dedsecn00b@gmail.com>