Jump from 2nd stage bootloader in Calypso OS and execute the __start entry point of the OS
 ## Summary of changes

Add statup.c source which contains the interrupt vector table and the
initial stack pointer value. The stack is configured to be 2048 for the
OS initialization part. Update the defconfig flash command with :

make load # Converts the .elf file into an uf2 file which can be dragged
and dropped on the RpiPico executing the ROM bootloader in USB mass
storage mode.

To execute the RpiPico ROM in USB Mass Storage mode simply plug the USB
cable while holding the BOOTSEL switch. Otherwise simply use a jtag to
flash the board on the SWD pins. Currently blinking a led works.

Signed-off-by: dedsecn00b <dedsecn00b@gmail.com>