rpipico: Remove the rpipico-sdk submodule
Avoid having a dependency on the external pico-sdk and
remove the submodule from arch/arm/rpipico/pico-sdk
Add CMSIS library as a submodule to get the
cm0plus core definitions.
Modify the linker script so that it will no longer
append the second stage bootloader generated by
pico-sdk.
At this point drivers(GPIO, UART) no longer rely on
the pico-sdk to drive the peripherals. The fix
for GPIO & UART will be added in a subsequent patch.

Signed-off-by: Sebastian Ene <sebastian.ene07@gmail.com>